Cleaning Code
Added more modularization to code by breaking up parts of the movie page into different sections stored within other classes
