+++ frontend/src/styles/Movie.module.css
@@ -182,99 +182,6 @@
 
 }
 
-.movie-cast-and-crew-container {
-    height: 600px;
-    width: 100%;
-    display: flex;
-    justify-content: center;
-
-}
-
-/* .people .person-card:first-child {
-    margin-left: 0;
-}
-
-.people .person-card:last-child {
-    margin-right: 0;
-} */
-
-.movie-cast-and-crew {
-    margin-top: 50px;
-    width: 80vw;
-}
-
-.movie-person-scroller{
-    position: relative;
-    top: 0;
-    left: 0;
-
-}
-
-#tab-bar {
-    margin-bottom: 16px;
-}
-#tab-bar2 {
-    margin-bottom: 16px;
-}
-
-ol.people {
-    list-style-type: none;
-    list-style-position: inside;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    position: relative;
-    top: 0;
-    left: 0;
-}
-
-ol.people {
-    -webkit-overflow-scrolling: touch;
-    overflow-y: hidden;
-    overflow-x: scroll;
-    margin-left: -10px;
-    margin-top: -10px;
-    padding-bottom: 10px
-}
-
-li.person-card {
-    margin: 10px 12px 10px 12px;
-    border: 1px solid rgba(var(--lightGrey), 1);
-    padding-bottom: 10px;
-    border-radius: 8px;
-    overflow: hidden;
-    min-width: 180px;
-    max-width: 180px;
-    background-color: white;
-    box-shadow: 0 2px 8px rgba(255,255,255,0.7);
-    color: black;
-    justify-content: center;
-    /* height: 225px; */
-}
-
-.person-image {
-    width: 100%;
-    height: 270px;
-
-}
-.person-info {
-    /* display: flex; */
-    /* justify-content: center; */
-    margin-top: 10px;
-    text-align: center;
-}
-
-.person-name { 
-    font-size: 18px;
-    font-weight: 700;
-    margin: 0;
-}
-
-.person-role {
-
-}
-
-
 
 .review-container {
     height: auto;
@@ -482,97 +389,3 @@ li.person-card {
     
 
 }
-
-.movie-media-menu {
-    width: 80vw;
-}
-
-.media-scroller{
-    position: relative;
-    top: 0;
-    left: 0;
-    margin-bottom: 100px;
-}
-
-ol.media {
-    list-style-type: none;
-    list-style-position: inside;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    position: relative;
-    top: 0;
-    left: 0;
-    align-items: center;
-}
-
-ol.media {
-    -webkit-overflow-scrolling: touch;
-    overflow-y: hidden;
-    overflow-x: scroll;
-    margin-left: -10px;
-    margin-top: -10px;
-    padding-bottom: 10px
-}
-
-li.poster-card {
-    margin: 10px 12px 10px 12px;
-    border: 1px solid rgba(var(--lightGrey), 1);
-    border-radius: 8px;
-    overflow: hidden;
-    min-width: 180px;
-    max-width: 180px;
-    background-color: black;
-    box-shadow: 0 2px 8px rgba(255,255,255,0.7);
-    color: black;
-    justify-content: center;
-    /* height: 225px; */
-}
-
-.poster-image {
-    width: 100%;
-    /* height: 270px; */
-
-}
-
-li.backdrop-card {
-    margin: 10px 12px 10px 12px;
-    border: 1px solid rgba(var(--lightGrey), 1);
-    border-radius: 8px;
-    overflow: hidden;
-    min-width: 270px;
-    max-width: 270px;
-    background-color: black;
-    box-shadow: 0 2px 8px rgba(255,255,255,0.7);
-    color: black;
-    justify-content: center;
-    /* height: 225px; */
-}
-
-.backdrop-image {
-    width: 100%;
-    /* height: 270px; */
-
-}
-
-li.logo-card {
-    margin: 10px 12px 10px 12px;
-    border: 1px solid rgba(var(--lightGrey), 1);
-    border-radius: 8px;
-    overflow: hidden;
-    min-width: 270px;
-    max-width: 270px;
-    background-color: black;
-    box-shadow: 0 2px 8px rgba(255,255,255,0.7);
-    color: black;
-    justify-content: center;
-    /* height: 225px; */
-}
-
-.logo-image {
-    width: 100%;
-    /* height: 270px; */
-
-}
-
-
