+++ frontend/src/styles/MovieMedia.module.css
@@ -1,0 +1,105 @@
+.movie-media {
+    height: auto;
+    /* width: 100%; */
+    /* display: flex;
+    justify-content: center; */
+    /* height: 200px; */
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    
+
+}
+
+.movie-media-menu {
+    width: 80vw;
+}
+
+.media-scroller{
+    position: relative;
+    top: 0;
+    left: 0;
+    margin-bottom: 100px;
+}
+
+ol.media {
+    list-style-type: none;
+    list-style-position: inside;
+    margin: 0;
+    padding: 0;
+    display: flex;
+    position: relative;
+    top: 0;
+    left: 0;
+    align-items: center;
+}
+
+ol.media {
+    -webkit-overflow-scrolling: touch;
+    overflow-y: hidden;
+    overflow-x: scroll;
+    margin-left: -10px;
+    margin-top: -10px;
+    padding-bottom: 10px
+}
+
+li.poster-card {
+    margin: 10px 12px 10px 12px;
+    border: 1px solid rgba(var(--lightGrey), 1);
+    border-radius: 8px;
+    overflow: hidden;
+    min-width: 180px;
+    max-width: 180px;
+    background-color: black;
+    box-shadow: 0 2px 8px rgba(255,255,255,0.7);
+    color: black;
+    justify-content: center;
+    /* height: 225px; */
+}
+
+.poster-image {
+    width: 100%;
+    /* height: 270px; */
+
+}
+
+li.backdrop-card {
+    margin: 10px 12px 10px 12px;
+    border: 1px solid rgba(var(--lightGrey), 1);
+    border-radius: 8px;
+    overflow: hidden;
+    min-width: 270px;
+    max-width: 270px;
+    background-color: black;
+    box-shadow: 0 2px 8px rgba(255,255,255,0.7);
+    color: black;
+    justify-content: center;
+    /* height: 225px; */
+}
+
+.backdrop-image {
+    width: 100%;
+    /* height: 270px; */
+
+}
+
+li.logo-card {
+    margin: 10px 12px 10px 12px;
+    border: 1px solid rgba(var(--lightGrey), 1);
+    border-radius: 8px;
+    overflow: hidden;
+    min-width: 270px;
+    max-width: 270px;
+    background-color: black;
+    box-shadow: 0 2px 8px rgba(255,255,255,0.7);
+    color: black;
+    justify-content: center;
+    /* height: 225px; */
+}
+
+.logo-image {
+    width: 100%;
+    /* height: 270px; */
+
+}
+
